Trait mpesa::mpesa_security::MpesaSecurity

pub trait MpesaSecurity {
    fn gen_security_credentials(&self) -> Result<String, Box<dyn Error>>;
}

Trait responsible for implementation of security configs for Mpesa

Required methods

fn gen_security_credentials(&self) -> Result<String, Box<dyn Error>>

Generates security credentials M-Pesa Core authenticates a transaction by decrypting the security credentials. Security credentials are generated by encrypting the base64 encoded initiator password with M-Pesa’s public key, a X509 certificate. Returns base64 encoded string.

Loading content...