Module mpesa::payloads

Structs

AccountBalancePayload

AccountBalanceResponse
B2C response Field names deliberately in Pascal case to correctly deserialize the response data

AuthResponse
Response from calling the Safaricom OAuth endpoint

B2bPayload

B2bResponse
B2C response Field names deliberately in Pascal case to correctly deserialize the response data

B2cPayload
Payload to allow for b2c transactions: See https://developer.safaricom.co.ke/docs#b2c-api for a detailed description of each field.

B2cResponse
B2C response Field names deliberately in Pascal case to correctly deserialize the response data

C2bRegisterPayload
Payload to register the 3rd party’s confirmation and validation URLs to M-Pesa See more here: https://developer.safaricom.co.ke/docs?shell#c2b-api

C2bRegisterResponse
C2B register response Field names deliberately in Pascal case to correctly deserialize the response data

C2bSimulatePayload
Payload to make payment requests from C2B. See more: https://developer.safaricom.co.ke/docs#c2b-api

C2bSimulateResponse
C2B payment response

Enums

ResponseType
C2B Register Response types